temporary g-aio update please wait

+++ PySoft/GUI/G-AIO.py
@@ -97,12 +97,14 @@ class WinAIO(Frame):
             wordchoice = "uninstall G-AIO from your computer."
         if parameter == 2:
             a = messagebox.showinfo("WinAIO Setup Information","WinAIO is a plug-in for Windows that replaces the standard shell with this program. It changes the Shell registry key and install itself in C:\\WinAIO folder. This install directory cannot be changed.\n\nNOTE: To use this function of this program, you must agree to the license and its terms of conditions as said in the GitHub GamerSoft24/Software repository, as well as any and all disclaimers and warranty information (if applicable). We will not be liable for any damages caused by any file, software package, individual program or other material from this repository in your possession (this includes, but is not limited to, modification, execution or download of the files)! This includes, but is not limited to, unintentional bugs, user error caused by an unclear prompt, clearly marked dangerous programs that may crash your computer or user negligence (didn't read the warnings) and more.")
+            return
         if os.name != 'nt':
             a = messagebox.showerror("The OS that you are using is not compatible with WinAIO Software. ")
         a = messagebox.showinfo("WinAIO Setup",f"Setup will now {wordchoice}")
         if parameter == 0:
             a = messagebox.askyesno("WinAIO Setup","Setup will override the default Windows menu that you may be used to and replace it with this program. The developers of this program are in no way, shape or form liable for any data loss that occurs, to the extent of the LICENSE and warranty agreements in the GitHub GamerSoft24/Software repository.\nDo you with to continue?",icon=messagebox.WARNING)
             if a:
+                os.system("md C:\\WinAIO")
                 with open("C:\\WinAIO\\SETUP.BAT", "w") as authadmin:
                     authadmin.write(r"@echo off")
                     authadmin.write(r"net session >nul 2>&1")
@@ -114,7 +116,9 @@ class WinAIO(Frame):
                 os.system("C:\\WinAIO\\SETUP.BAT")
             else:
                 a = messagebox.showerror("WinAIO Setup","The operation was cancelled.")
+                return
         if parameter == 1:       
+                os.system("md C:\\WinAIO")
                 with open("C:\\WinAIO\\UNSETUP.BAT", "w") as authadmin:
                     authadmin.write(r"@echo off")
                     authadmin.write(r"net session >nul 2>&1")
@@ -1403,20 +1407,21 @@ class PassManager(Frame):
         x = messagebox.showinfo("G-AIO","The password has generated and saved successfully")
         return
 if __name__ == "__main__":#uarte
-    if sys.argv[1] == "--setup":
-        a = os.system("reg add HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\System /v VerboseStatus /t REG_DWORD /d 1 /f")
-        b = os.system(f"copy {os.path.abspath(__file__)} C:\\WinAIO\\G-AIO.PY")
-        c = os.system("reg add HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon /v Shell /t REG_SZ /d C:\WinAIO\G-AIO.PY /f")
-        if a != 0 or b != 0 or c != 0:
-            d = messagebox.showerror("WinAIO Setup","An error has occured while configuring your computer. Please see the terminal logs for more information (usually the black window with text)")
-        d = messagebox.askyesno("WinAIO Setup","Setup has finished configuring your computer. For the changes to take effect, you must restart your computer. Do you wish to do that now?")
-        if d:
-            os.system("shutdown /r /t 0 ")
-    if sys.argv[1] == "--unsetup":
-        a = os.system("reg add HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon /v Shell /t REG_SZ /d C:\WINDOWS\EXPLORER.EXE /f")
-        d = messagebox.askyesno("WinAIO Setup","Setup has finished configuring your computer. For the changes to take effect, you must restart your computer. Do you wish to do that now?")
-        if d:
-            os.system("shutdown /r /t 0 ")
+    if len(sys.argv) > 1:
+        if sys.argv[1] == "--setup":
+            a = os.system("reg add HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\System /v VerboseStatus /t REG_DWORD /d 1 /f")
+            b = os.system(f"copy {os.path.abspath(__file__)} C:\\WinAIO\\G-AIO.PY")
+            c = os.system("reg add HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon /v Shell /t REG_SZ /d C:\WinAIO\G-AIO.PY /f")
+            if a != 0 or b != 0 or c != 0:
+                d = messagebox.showerror("WinAIO Setup","An error has occured while configuring your computer. Please see the terminal logs for more information (usually the black window with text)")
+            d = messagebox.askyesno("WinAIO Setup","Setup has finished configuring your computer. For the changes to take effect, you must restart your computer. Do you wish to do that now?")
+            if d:
+                os.system("shutdown /r /t 0 ")
+        if sys.argv[1] == "--unsetup":
+            a = os.system("reg add HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon /v Shell /t REG_SZ /d C:\WINDOWS\EXPLORER.EXE /f")
+            d = messagebox.askyesno("WinAIO Setup","Setup has finished configuring your computer. For the changes to take effect, you must restart your computer. Do you wish to do that now?")
+            if d:
+                os.system("shutdown /r /t 0 ")
     app = AssistantApp()
     app.geometry("800x600")
     app.resizable(width=False,height=False)
